+++ PaymentManagement/PaymentManagementCore/src/main/java/com/warehouse/paypal/domain/model/PaymentRequest.java
@@ -2,9 +2,11 @@ package com.warehouse.paypal.domain.model;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class PaymentRequest {
     Long parcelId;
-    double price;
+    BigDecimal price;
 }
